% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulapdf-package.R
\docType{package}
\name{tabulapdf}
\alias{tabulapdf}
\title{Java '.jar' files of 'Tabula' PDF table extractor library.}
\description{
'Tabula PDF' is a Java library designed to computationally extract
tables from PDF documents.
}
\references{
See \href{https://tabula.technology/}{Tabula} webpage and 
\href{https://github.com/ropensci/tabulapdf/}{tabulapdf} package for 
more details and all the features.
}
